Remove per-instance optional state for "Use AYON Entity URI"
Due to a limitation of the Publisher UI the attribute definitions for the `usdrop` family won't be shown in the Publisher UI. The Publisher UI only shows attribute definitions for the main primary product type, not for the other "families" on the Created Instance. Because `usd` is the primary product type, and not `usdrop` the attributes would never show.