Support 'hero' and 'latest' in AYON HDAs version parameters
Also, if `use_entity_uri` is enabled on the node do not provide the "resolved" filepath but set the filepath using the actual constructed AYON Entity URI of the inputs parameters, so that e.g. 'latest' does not turn into a path with version=1 but remains version=latest in the path parameter.